rag_java_claude/pyproject.toml
[project]
name = "java-qa-agent"
version = "0.1.0"
description = "Java code Q&A RAG agent"
requires-python = ">=3.10"
dependencies = [
    "langchain-community>=0.3.0",
    "chromadb>=0.5.0",
    "ollama>=0.3.0",
    "typer>=0.12.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
]

[project.scripts]
java-qa = "java_qa_agent.cli:app"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.8.0",
    "ruff>=0.4.0",
    "types-PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/java_qa_agent"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
